# [DEBUG-4070] dyninst/object: use faster zlib